+++ RickAndMorty/RickAndMorty/Managers/RMImageLoader.swift
@@ -1,0 +1,42 @@
+//
+//  RMImageLoader.swift
+//  RickAndMorty
+//
+//  Created by Ramazan Abdulaev on 07.01.2023.
+//
+
+import Foundation
+
+final class RMImageLoader {
+    
+    static let shared = RMImageLoader()
+    
+    private var imageDataCache = NSCache<NSString, NSData>()
+    
+    private init() {}
+    
+    /// Get image content with url
+    /// - Parameters:
+    ///   - url: source url
+    ///   - completion: callback
+    public func downloadImage(url: URL, completion: @escaping (Result<Data, Error>) -> Void) {
+        let key = url.absoluteString as NSString
+        if let data = imageDataCache.object(forKey: key) {
+            completion(.success(data as Data))
+            return
+        }
+        
+        let request = URLRequest(url: url)
+        let task = URLSession.shared.dataTask(with: request) { [weak self] data, _, error in
+            guard let data, error == nil else {
+                completion(.failure(error ?? URLError(.badServerResponse)))
+                return
+            }
+            let key = url.absoluteString as NSString
+            let object = data as NSData
+            self?.imageDataCache.setObject(object, forKey: key)
+            completion(.success(data))
+        }
+        task.resume()
+    }
+}
