+++ RickAndMorty/RickAndMorty/ViewModels/RMCharacterCollectionViewCellViewModel.swift
@@ -39,14 +39,6 @@ final class RMCharacterCollectionViewCellViewModel {
             completion(.failure(URLError(.badURL)))
             return
         }
-        let request = URLRequest(url: url)
-        let task = URLSession.shared.dataTask(with: request) { data, _, error in
-            guard let data, error == nil else {
-                completion(.failure(error ?? URLError(.badServerResponse)))
-                return
-            }
-            completion(.success(data))
-        }
-        task.resume()
+        RMImageLoader.shared.downloadImage(url: url, completion: completion)
     }
 }
